+++ packages/base/sync/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `rawChanges` array to `push` function parameters to make all changes available during push.
+* Added `modifiedFields` to `changes` object on `push` to allow patch updates
 
 ## [1.2.2] - 2025-03-20